+++ packages/functions/tsdown.config.ts
@@ -5,7 +5,7 @@ export default defineConfig({
 	bundle: false,
 	clean: true,
 	dts: true,
-	entry: ["src/index.ts", "src/v1.ts", "src/lib/**/*.ts", "!src/**/*.test.*"],
+	entry: ["src/index.ts", "src/lib/**/*.ts", "!src/**/*.test.*"],
 	format: "esm",
 	outDir: "dist",
 	treeshake: true,